requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
watchdog==3.0.0
requests==2.31.0
numpy==1.24.3
torch==2.1.0
torchaudio==2.1.0
speechbrain==0.5.15
soundfile==0.12.1
reportlab==4.0.7
pypdf>=3.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=10.0.0
dropbox>=11.36.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
boxsdk>=3.9.0
cryptography>=41.0.0